certService/src/main/java/org/onap/oom/certservice/certification/UpdateRequestTypeDetector.java
/*
 * ============LICENSE_START=======================================================
 * Cert Service
 * ================================================================================
 * Copyright (C) 2021 Nokia. All rights reserved.
 * ================================================================================
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 * ============LICENSE_END=========================================================
 */

package org.onap.oom.certservice.certification;

import org.onap.oom.certservice.certification.model.CertificateData;
import org.springframework.stereotype.Service;

@Service
public class UpdateRequestTypeDetector {

    public boolean isKur(CertificateData certificateData1, CertificateData certificateData2) {
        return certificateData1.equals(certificateData2);
    }

}
